+++ Sources/MacParakeetCore/Services/Telemetry/TelemetryEvent.swift
@@ -502,6 +502,7 @@ public enum TelemetrySettingName: String, Sendable, Equatable {
     case meetingAudioRetention = "meeting_audio_retention"
     case youtubeAudioQuality = "youtube_audio_quality"
     case speakerDiarization = "speaker_diarization"
+    case meetingSpeakerDiarization = "meeting_speaker_diarization"
     case parakeetModelVariant = "parakeet_model_variant"
     case nemotronModelVariant = "nemotron_model_variant"
     case whisperDefaultLanguage = "whisper_default_language"